Organizando e refatorando Lexer.
