+++ PasC/PasC/PasC.csproj
@@ -48,47 +48,7 @@
     <Compile Include="Models\Tag.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="States\Lexer.cs" />
-    <Compile Include="States\State00.cs" />
-    <Compile Include="States\State01.cs" />
-    <Compile Include="States\State10.cs" />
-    <Compile Include="States\State11.cs" />
-    <Compile Include="States\State12.cs" />
-    <Compile Include="States\State13.cs" />
-    <Compile Include="States\State14.cs" />
-    <Compile Include="States\State15.cs" />
-    <Compile Include="States\State16.cs" />
-    <Compile Include="States\State17.cs" />
-    <Compile Include="States\State18.cs" />
-    <Compile Include="States\State19.cs" />
-    <Compile Include="States\State02.cs" />
-    <Compile Include="States\State20.cs" />
-    <Compile Include="States\State21.cs" />
-    <Compile Include="States\State22.cs" />
-    <Compile Include="States\State23.cs" />
-    <Compile Include="States\State24.cs" />
-    <Compile Include="States\State25.cs" />
-    <Compile Include="States\State26.cs" />
-    <Compile Include="States\State27.cs" />
-    <Compile Include="States\State28.cs" />
-    <Compile Include="States\State29.cs" />
-    <Compile Include="States\State03.cs" />
-    <Compile Include="States\State30.cs" />
-    <Compile Include="States\State32.cs" />
-    <Compile Include="States\State33.cs" />
-    <Compile Include="States\State34.cs" />
-    <Compile Include="States\State35.cs" />
-    <Compile Include="States\State36.cs" />
-    <Compile Include="States\State38.cs" />
-    <Compile Include="States\State39.cs" />
-    <Compile Include="States\State37.cs" />
-    <Compile Include="States\State31.cs" />
-    <Compile Include="States\State04.cs" />
-    <Compile Include="States\State05.cs" />
-    <Compile Include="States\State06.cs" />
-    <Compile Include="States\State07.cs" />
-    <Compile Include="States\State08.cs" />
-    <Compile Include="States\State09.cs" />
+    <Compile Include="Modules\Lexer.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
